refactor: the `data` option was renamed to the `prependData` option
BREAKING CHANGE: the `data` option was renamed to the `prependData` option
