+++ test/loader.test.js
@@ -109,42 +109,6 @@ describe('loader', () => {
         expect(stats.compilation.errors).toMatchSnapshot('errors');
       });
 
-      it(`should work with the "data" option as a string (${implementationName}) (${syntax})`, async () => {
-        const testId = getTestId('prepending-data', syntax);
-        const options = {
-          implementation: getImplementationByName(implementationName),
-          data: `$prepended-data: hotpink${syntax === 'sass' ? '\n' : ';'}`,
-        };
-        const stats = await compile(testId, { loader: { options } });
-
-        expect(getCodeFromBundle(stats).css).toBe(
-          getCodeFromSass(testId, options).css
-        );
-
-        expect(stats.compilation.warnings).toMatchSnapshot('warnings');
-        expect(stats.compilation.errors).toMatchSnapshot('errors');
-      });
-
-      it(`should work with the "data" option as a function (${implementationName}) (${syntax})`, async () => {
-        const testId = getTestId('prepending-data', syntax);
-        const options = {
-          implementation: getImplementationByName(implementationName),
-          data: (loaderContext) => {
-            expect(loaderContext).toBeDefined();
-
-            return `$prepended-data: hotpink${syntax === 'sass' ? '\n' : ';'}`;
-          },
-        };
-        const stats = await compile(testId, { loader: { options } });
-
-        expect(getCodeFromBundle(stats).css).toBe(
-          getCodeFromSass(testId, options).css
-        );
-
-        expect(stats.compilation.warnings).toMatchSnapshot('warnings');
-        expect(stats.compilation.errors).toMatchSnapshot('errors');
-      });
-
       it(`should work with the "includePaths" option (${implementationName}) (${syntax})`, async () => {
         const testId = getTestId('import-include-paths', syntax);
         const options = {
